Switch IAmountUI to a separate file for better organization.
Moved the IAmountUI interface out of AmountUI.cs and into its own file, IAmountUI.cs. This change improves code structure and keeps interfaces and classes separate for clarity and maintainability.
